+++ app/src/main/java/com/example/a16_room/data/repository/SubjectRepository.kt
@@ -13,7 +13,7 @@ class SubjectRepository(context: Context) {
         return subjectDAO.insert(subject)
     }
 
-    fun insertStudentSubject(studentId: Int, subjectId: Int) {
+    fun insertStudentSubject(studentId: Long, subjectId: Int) {
         val studentSubjectCrossRef = StudentSubjectCrossRef(studentId, subjectId)
         subjectDAO.insertStudentSubjectCrossRef(studentSubjectCrossRef)
     }
